// src/types/schemas/wardrobeSchemas.ts
export type FurnitureTableSchema = {
    id: number
    wardrobeId: number
    specId: number
    minWidth: number
    maxWidth: number
    minHeight: number
    maxHeight: number
    minDepth: number
    maxDepth: number
    count: number
    size: string
    charId: number
}